requirements.txt
Flask==0.12.1
azure-storage==0.34.3

